+++ src/Services/IM/Notify/Result/NotifySchemaItemResult.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\IM\Notify\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractAnnotatedItem;
+
+/**
+ * @property-read string $name
+ * @property-read string $module_id
+ * @property-read array $list
+ */
+class NotifySchemaItemResult extends AbstractAnnotatedItem
+{
+}